User thymeleaf controllers some methods, Like and Comments tested, Theme added

+++ src/main/java/com/example/demo/services/CommentServiceImpl.java
@@ -1,6 +1,9 @@
 package com.example.demo.services;
 
+import com.example.demo.exceptions.AuthorizationException;
+import com.example.demo.exceptions.EntityNotFoundException;
 import com.example.demo.models.Comments;
+import com.example.demo.models.User;
 import com.example.demo.repositories.CommentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,15 +14,17 @@ import java.util.List;
 public class CommentServiceImpl implements CommentService {
 
     private CommentRepository commentRepository;
+    private UserService userService;
 
     @Autowired
-    public CommentServiceImpl(CommentRepository commentRepository) {
+    public CommentServiceImpl(CommentRepository commentRepository, UserService userService) {
         this.commentRepository = commentRepository;
+        this.userService = userService;
     }
 
     @Override
     public Comments getById(int commentId) {
-        return commentRepository.getById(commentId);
+        return commentRepository.getByCommentId(commentId);
     }
 
     @Override
@@ -34,13 +39,24 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
+    public List<Comments> getCommentsByUserId(int userId) {
+        return commentRepository.getCommentsByUserId(userId);
+    }
+
+    @Override
     public Comments updateComment(Comments comments) {
         commentRepository.save(comments);
         return comments;
     }
 
     @Override
-    public void deleteComment(int id) {
-        commentRepository.deleteById(id);
+    public void deleteComment(int id, String username) {
+        User user = userService.getByUsername(username);
+        List<Comments> userComments = getCommentsByUserId(user.getId());
+        if (userComments.contains(commentRepository.getByCommentId(id))) {
+            commentRepository.deleteById(id);
+        } else {
+            throw new AuthorizationException("You have not permissions to delete this comment");
+        }
     }
 }
